Fix X / Y position flags in documentation.
I tried to use the code included in the README to put together a
quick POC app and noticed the "anchor_x_offset" and
"anchor_y_offset" flags didn't seem to change the location of the
signature, always ending up in the upper-right hand corner.

After some digging, I found that the client.rb expects this data to
be passed in as "x_position" and "y_position".  After updating the
POC to use these keys, the signature would move around as expected.

I've updated the documentation to include the correct data flags.
This should allow new users to get up to speed a lot quicker, without
having to alter a local copy of the library to find the right flags.