Record base_url returned from successful login
Per Docusign team: "In the demo environment all account URLs
start with demo.docusign.net, however in production there are
multiple sub-domains like www, na2, eu, etc.  To resolve all you need to
do is parse the baseUrl value that is returned from the login call."

This seems to also apply to the oauth token, but I'm not sure of the use
case there.

Thanks to Will Beattie for the report.