Do not assign clientuserId unless embedded
Previously this defaulted to always sending the clientuserId to
docusign. At which point docusing will assume this document is to be
signed embedded and will not send an email notification to the signer
via email. By setting it if the the embedded key is provided we can use
this in both instances.